[API] Moves for_each_type to detail core and remove original header.
